refactor: visualization and utility
