+++ calibrators/marker_radar_lidar_calibrator/CMakeLists.txt
@@ -12,6 +12,8 @@ ament_python_install_package(${PROJECT_NAME})
 ament_auto_add_executable(marker_radar_lidar_calibrator
   src/marker_radar_lidar_calibrator.cpp
   src/transformation_estimator.cpp
+  src/visualization.cpp
+  src/utils.cpp
   src/main.cpp
 )
